+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Support for specifying view and view template at the same time with Page::setView().
 
+### Changed
+- Wireframe module now updates the internal settings hash more consistently. This is primarily used for runtime page render cache.
+
 ## [0.15.2] - 2020-09-13
 
 ### Changed